+++ frontend/app/src/main/java/com/apptive/japkor/ui/signup/SignUpScreen.kt
@@ -49,6 +49,8 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.focus.onFocusChanged
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.foundation.layout.imePadding
+import com.apptive.japkor.ui.signup.components.EmailWithAuthSection
+import com.apptive.japkor.ui.signup.components.PasswordSection
 
 
 /**
@@ -229,289 +231,3 @@ fun SignUpScreen(navController: NavController) {
     }
 }
 
-/**
- * 이메일 입력 + 인증 코드 전송/인증
- */
-@Composable
-private fun EmailWithAuthSection(
-    emailLocal: String,
-    onEmailLocalChange: (String) -> Unit,
-    emailDomain: String,
-    onEmailDomainChange: (String) -> Unit,
-    authCode: String,
-    onAuthCodeChange: (String) -> Unit,
-    canSendCode: Boolean,
-    onClickSendCode: () -> Unit,
-    canVerifyCode: Boolean,
-    onClickVerify: () -> Unit
-) {
-    Column(
-        modifier = Modifier.fillMaxWidth(),
-        verticalArrangement = Arrangement.spacedBy(8.dp)
-    ) {
-        // 라벨
-        CustomText(
-            text = "이메일",
-            type = CustomTextType.body,
-            color = CustomColor.black,
-            size = 15.sp
-        )
-
-        // 이메일 앞부분 + 도메인 (둘 다 직접 입력)
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            // 이메일 앞부분
-            HalfCustomTextField(
-                value = emailLocal,
-                onValueChange = onEmailLocalChange,
-                placeholder = "이메일",
-                modifier = Modifier.weight(1f),
-                keyboardOptions = KeyboardOptions(
-                    keyboardType = KeyboardType.Text,
-                    imeAction = ImeAction.Next
-                )
-            )
-            CustomText(
-                text = " @ ",
-                type = CustomTextType.body,
-                color = CustomColor.gray300,
-                modifier = Modifier.padding(horizontal = 6.dp)
-            )
-            // 도메인 직접 입력
-            HalfCustomTextField(
-                value = emailDomain,
-                onValueChange = onEmailDomainChange,
-                placeholder = "직접 입력",
-                modifier = Modifier.weight(1f),
-                keyboardOptions = KeyboardOptions(
-                    keyboardType = KeyboardType.Text,
-                    imeAction = ImeAction.Next
-                )
-            )
-        }
-
-        // 인증 코드 전송 버튼 (비활성/활성)
-        Button(
-            onClick = onClickSendCode,
-            enabled = canSendCode,
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(50.dp),
-            colors = ButtonDefaults.buttonColors(
-                containerColor = Color(0xFFF45C4A),
-                contentColor = CustomColor.white,
-                disabledContainerColor = CustomColor.gray300,
-                disabledContentColor = CustomColor.white
-            ),
-            shape = RoundedCornerShape(16.dp)
-        ) {
-            CustomText(
-                text = "인증 코드 전송",
-                type = CustomTextType.body,
-                size = 15.sp
-            )
-        }
-
-        // 인증 코드 입력 + 필드 안쪽에 인증 버튼
-        AuthCodeField(
-            authCode = authCode,
-            onAuthCodeChange = onAuthCodeChange,
-            canVerifyCode = canVerifyCode,
-            onClickVerify = onClickVerify
-        )
-    }
-}
-
-
-
-/**
- * CustomTextField 반으로 쪼갠 버전 (이메일 입력에서만 사용)
- */
-@Composable
-private fun HalfCustomTextField(
-    value: String,
-    onValueChange: (String) -> Unit,
-    placeholder: String,
-    modifier: Modifier = Modifier,
-    keyboardOptions: KeyboardOptions = KeyboardOptions.Default,
-    visualTransformation: VisualTransformation = VisualTransformation.None,
-) {
-    val shape = RoundedCornerShape(16.dp)
-    var isFocused by remember { mutableStateOf(false) }
-    val borderWidth = if (isFocused) 2.dp else 1.dp
-    val borderColor = if (isFocused) CustomColor.gray300 else CustomColor.gray200
-
-    Box(
-        modifier = modifier
-            .height(50.dp)
-            .border(borderWidth, borderColor, shape)
-            .background(CustomColor.white, shape)
-            .padding(horizontal = 16.dp),
-        contentAlignment = Alignment.CenterStart
-    ) {
-        if (value.isEmpty()) {
-            CustomText(
-                text = placeholder,
-                type = CustomTextType.body,
-                color = CustomColor.gray300,
-                size = 14.sp
-            )
-        }
-
-        BasicTextField(
-            value = value,
-            onValueChange = onValueChange,
-            singleLine = true,
-            textStyle = TextStyle(
-                color = CustomColor.black,
-                fontSize = 14.sp,
-                textAlign = TextAlign.Start
-            ),
-            modifier = Modifier
-                .fillMaxWidth()
-                .onFocusChanged { focusState ->
-                    isFocused = focusState.isFocused // 포커스 상태 업데이트 (never read라고 뜨지만 필요함)
-                },
-            keyboardOptions = keyboardOptions,
-            visualTransformation = visualTransformation
-        )
-    }
-}
-
-/**
- * 인증 코드 입력 필드
- */
-@Composable
-private fun AuthCodeField(
-    authCode: String,
-    onAuthCodeChange: (String) -> Unit,
-    canVerifyCode: Boolean,
-    onClickVerify: () -> Unit
-) {
-    val outerShape = RoundedCornerShape(16.dp)
-    var isFocused by remember { mutableStateOf(false) }
-    val borderWidth = if (isFocused) 2.dp else 1.dp
-    val borderColor = if (isFocused) CustomColor.gray300 else CustomColor.gray200
-
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(50.dp)
-            .border(borderWidth, borderColor, outerShape)
-            .background(Color.White, outerShape)
-            .padding(horizontal = 12.dp)
-    ) {
-        Row(
-            modifier = Modifier.fillMaxSize(),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            // 코드 입력 영역
-            Box(
-                modifier = Modifier
-                    .weight(1f)
-                    .padding(end = 8.dp),
-                contentAlignment = Alignment.CenterStart
-            ) {
-                if (authCode.isEmpty()) {
-                    CustomText(
-                        text = "인증 코드",
-                        type = CustomTextType.body,
-                        color = CustomColor.gray300,
-                        size = 14.sp
-                    )
-                }
-                BasicTextField(
-                    value = authCode,
-                    onValueChange = onAuthCodeChange,
-                    singleLine = true,
-                    textStyle = TextStyle(
-                        color = CustomColor.black,
-                        fontSize = 14.sp
-                    ),
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .onFocusChanged { focusState ->
-                            isFocused = focusState.isFocused
-                        }
-                )
-            }
-
-            // 필드 안에 들어가는 작은 인증 버튼
-            Button(
-                onClick = onClickVerify,
-                enabled = canVerifyCode,
-                modifier = Modifier.height(32.dp),
-                colors = ButtonDefaults.buttonColors(
-                    containerColor = Color(0xFFF45C4A),
-                    contentColor = Color.White,
-                    disabledContainerColor = CustomColor.gray300,
-                    disabledContentColor = Color.White
-                ),
-                shape = RoundedCornerShape(16.dp)
-            ) {
-                CustomText(
-                    text = "인증",
-                    type = CustomTextType.body,
-                    size = 12.sp
-                )
-            }
-        }
-    }
-}
-
-/**
- * 비밀번호 입력 및 확인
- */
-@Composable
-private fun PasswordSection(
-    password: String,
-    passwordConfirm: String,
-    onPasswordChange: (String) -> Unit,
-    onPasswordConfirmChange: (String) -> Unit,
-    isPasswordMismatch: Boolean
-) {
-    Column(
-        modifier = Modifier.fillMaxWidth(),
-        verticalArrangement = Arrangement.spacedBy(8.dp)
-    ) {
-        CustomText(
-            text = "비밀번호",
-            type = CustomTextType.body,
-            color = CustomColor.black,
-            size = 15.sp
-        )
-        CustomOutlinedTextField(
-            value = password,
-            onValueChange = onPasswordChange,
-            placeholder = "비밀번호",
-            isPassword = true
-        )
-        CustomOutlinedTextField(
-            value = passwordConfirm,
-            onValueChange = onPasswordConfirmChange,
-            placeholder = "비밀번호 확인",
-            isPassword = true
-        )
-
-        // 비밀번호 불일치 에러 메시지
-        if (isPasswordMismatch) {
-            CustomText(
-                text = "비밀번호가 일치하지 않습니다.",
-                color = Color(0xFFF45C4A),
-                type = CustomTextType.body,
-                modifier = Modifier.padding(horizontal = 4.dp),
-                size = 13.sp
-            )
-        }
-
-        CustomText(
-            text = "8 ~ 16자 이내 영문, 특수 문자 조합",
-            color = CustomColor.gray300,
-            type = CustomTextType.body,
-            modifier = Modifier.padding(horizontal = 4.dp),
-            size = 14.sp
-        )
-    }
-}
