+++ frontend/app/src/main/java/com/apptive/japkor/ui/signup/components/HalfCustomTextField.kt
@@ -1,0 +1,78 @@
+package com.apptive.japkor.ui.signup.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.BasicTextField
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.focus.onFocusChanged
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.input.VisualTransformation
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.apptive.japkor.ui.components.CustomText
+import com.apptive.japkor.ui.components.CustomTextType
+import com.apptive.japkor.ui.theme.CustomColor
+
+@Composable
+fun HalfCustomTextField(
+    value: String,
+    onValueChange: (String) -> Unit,
+    placeholder: String,
+    modifier: Modifier = Modifier,
+    keyboardOptions: KeyboardOptions = KeyboardOptions.Default,
+    visualTransformation: VisualTransformation = VisualTransformation.None,
+) {
+    val shape = RoundedCornerShape(16.dp)
+    var isFocused by remember { mutableStateOf(false) }
+    val borderWidth = if (isFocused) 2.dp else 1.dp
+    val borderColor = if (isFocused) CustomColor.gray300 else CustomColor.gray200
+
+    Box(
+        modifier = modifier
+            .height(50.dp)
+            .border(borderWidth, borderColor, shape)
+            .background(CustomColor.white, shape)
+            .padding(horizontal = 16.dp),
+        contentAlignment = Alignment.CenterStart
+    ) {
+        if (value.isEmpty()) {
+            CustomText(
+                text = placeholder,
+                type = CustomTextType.body,
+                color = CustomColor.gray300,
+                size = 14.sp
+            )
+        }
+
+        BasicTextField(
+            value = value,
+            onValueChange = onValueChange,
+            singleLine = true,
+            textStyle = TextStyle(
+                color = CustomColor.black,
+                fontSize = 14.sp,
+                textAlign = TextAlign.Start
+            ),
+            modifier = Modifier
+                .fillMaxWidth()
+                .onFocusChanged { focusState ->
+                    isFocused = focusState.isFocused // 포커스 상태 업데이트 (never read라고 뜨지만 필요함)
+                },
+            keyboardOptions = keyboardOptions,
+            visualTransformation = visualTransformation
+        )
+    }
+}
